# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: NCKH SaaS Multi-tenancy API
  version: 0.1.0
  description: >-
    Hợp đồng REST cho control plane và application plane. Các API nghiệp vụ lấy
    tenant từ host và JWT; không nhận tenant_id từ payload để phân quyền.
servers:
  - url: http://accounts.localhost:8080/api/v1
    description: Control host local
  - url: http://{tenant}.localhost:8080/api/v1
    description: Tenant host local
    variables:
      tenant:
        default: pool-demo
security:
  - bearerAuth: []
...
paths:
...
  /tenants:
    get:
      tags: [Tenants]
      operationId: listMyTenants
      responses:
        '200':
          description: Active memberships of current user
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/TenantView'}
        default: {$ref: '#/components/responses/Error'}
    post:
      tags: [Tenants]
      operationId: createTenant
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateTenantRequest'}
      responses:
        '200':
          description: Tenant in PENDING_PAYMENT state
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TenantView'}
        default: {$ref: '#/components/responses/Error'}
...
  /admin/tenants:
    get:
      tags: [Administration]
      operationId: listAdminTenants
      parameters:
        - name: page
          in: query
          schema: {type: integer, minimum: 0, default: 0}
        - name: size
          in: query
          schema: {type: integer, minimum: 1, maximum: 100, default: 20}
        - name: search
          in: query
          schema: {type: string, default: ''}
        - name: status
          in: query
          schema: {$ref: '#/components/schemas/TenantStatus'}
        - name: placement
          in: query
          schema: {$ref: '#/components/schemas/TenantPlacement'}
      responses:
        '200':
          description: Control-plane tenant page for system administrators
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminTenantPage'}
        default: {$ref: '#/components/responses/Error'}
  /admin/tenants/{tenantId}:
    get:
      tags: [Administration]
      operationId: getAdminTenant
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: Control-plane tenant detail, latest payment and provisioning transition history
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminTenantDetailView'}
        default: {$ref: '#/components/responses/Error'}
  /admin/tenants/{tenantId}/provisioning/retry:
    post:
      tags: [Administration]
      operationId: retryProvisioning
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200': {description: Failed job queued again and transition audited}
        default: {$ref: '#/components/responses/Error'}
  /admin/tenants/{tenantId}/resource-dead-letters:
    get:
      tags: [Administration]
      operationId: listResourceDeadLetters
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: page
          in: query
          schema: {type: integer, minimum: 0, default: 0}
        - name: size
          in: query
          schema: {type: integer, minimum: 1, maximum: 100, default: 20}
      responses:
        '200':
          description: Tenant-scoped resource cleanup events that exhausted automatic retries
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ResourceDeadLetterPage'}
        default: {$ref: '#/components/responses/Error'}
  /admin/tenants/{tenantId}/resource-dead-letters/{eventId}/requeue:
    post:
      tags: [Administration]
      operationId: requeueResourceDeadLetter
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: eventId
          in: path
          required: true
          schema: {$ref: '#/components/schemas/UUID'}
      responses:
        '200': {description: Dead-lettered cleanup event reset for retry and action audited}
        default: {$ref: '#/components/responses/Error'}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    TenantId:
      name: tenantId
      in: path
      required: true
      schema: {$ref: '#/components/schemas/UUID'}
...
  responses:
    Error:
      description: Structured API error
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiError'}
  schemas:
    UUID: {type: string, format: uuid}
    Instant: {type: string, format: date-time}
    TenantRole: {type: string, enum: [OWNER, ADMIN, MEMBER]}
...
    TenantTier: {type: string, enum: [STARTER, PROFESSIONAL, ENTERPRISE]}
    TenantPlacement: {type: string, enum: [POOL, SILO_DATABASE]}
    TenantStatus: {type: string, enum: [PENDING_PAYMENT, PROVISIONING, ACTIVE, FAILED, SUSPENDED]}
    PaymentStatus: {type: string, enum: [CREATED, PENDING, SUCCEEDED, FAILED, EXPIRED]}
    ProvisioningStatus: {type: string, enum: [QUEUED, RUNNING, SUCCEEDED, RETRYABLE_FAILED, FAILED_ROLLED_BACK, ROLLBACK_FAILED]}
...
    ApiError:
      type: object
      required: [timestamp, status, code, message, path, fieldErrors]
      properties:
        timestamp: {$ref: '#/components/schemas/Instant'}
        status: {type: integer}
        code: {type: string}
        message: {type: string}
        path: {type: string}
        fieldErrors: {type: object, additionalProperties: {type: string}}
        requestId: {type: [string, 'null']}
...
    TenantView:
      type: object
      required: [id, slug, name, tier, status, placement, role]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        slug: {type: string}
        name: {type: string}
        tier: {$ref: '#/components/schemas/TenantTier'}
        status: {$ref: '#/components/schemas/TenantStatus'}
        placement: {$ref: '#/components/schemas/TenantPlacement'}
        role: {$ref: '#/components/schemas/TenantRole'}
...
    CreateTenantRequest:
      type: object
      required: [slug, name, tier, placement]
      properties:
        slug: {type: string, pattern: '^[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?$'}
        name: {type: string}
        tier: {$ref: '#/components/schemas/TenantTier'}
        placement: {$ref: '#/components/schemas/TenantPlacement'}
...
    AdminTenantView:
      type: object
      required: [id, name, slug, tier, status, memberCount, createdAt]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        name: {type: string}
        slug: {type: string}
        tier: {$ref: '#/components/schemas/TenantTier'}
        placement: {anyOf: [{$ref: '#/components/schemas/TenantPlacement'}, {type: 'null'}]}
        status: {$ref: '#/components/schemas/TenantStatus'}
        provisioningStatus: {anyOf: [{$ref: '#/components/schemas/ProvisioningStatus'}, {type: 'null'}]}
        memberCount: {type: integer, format: int64}
        createdAt: {$ref: '#/components/schemas/Instant'}
        lastError: {type: [string, 'null']}
        paymentStatus: {anyOf: [{$ref: '#/components/schemas/PaymentStatus'}, {type: 'null'}]}
        paymentProvider: {type: [string, 'null']}
    AdminPaymentView:
      type: object
      required: [id, provider, status, amountMinor, currency, createdAt]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        provider: {type: string}
        status: {$ref: '#/components/schemas/PaymentStatus'}
        amountMinor: {type: integer, format: int64}
        currency: {type: string}
        createdAt: {$ref: '#/components/schemas/Instant'}
    AdminProvisioningView:
      type: object
      required: [id, status, attempts, createdAt]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        status: {$ref: '#/components/schemas/ProvisioningStatus'}
        attempts: {type: integer}
        lastErrorCode: {type: [string, 'null']}
        lastErrorMessage: {type: [string, 'null']}
        nextAttemptAt: {anyOf: [{$ref: '#/components/schemas/Instant'}, {type: 'null'}]}
        createdAt: {$ref: '#/components/schemas/Instant'}
    ProvisioningEventView:
      type: object
      required: [id, toStatus, attempt, createdAt]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        fromStatus: {anyOf: [{$ref: '#/components/schemas/ProvisioningStatus'}, {type: 'null'}]}
        toStatus: {$ref: '#/components/schemas/ProvisioningStatus'}
        attempt: {type: integer}
        errorCode: {type: [string, 'null']}
        message: {type: [string, 'null']}
        createdAt: {$ref: '#/components/schemas/Instant'}
    AdminTenantDetailView:
      type: object
      required: [tenant, events]
      properties:
        tenant: {$ref: '#/components/schemas/AdminTenantView'}
        payment: {anyOf: [{$ref: '#/components/schemas/AdminPaymentView'}, {type: 'null'}]}
        provisioning: {anyOf: [{$ref: '#/components/schemas/AdminProvisioningView'}, {type: 'null'}]}
        events:
          type: array
          items: {$ref: '#/components/schemas/ProvisioningEventView'}
    AdminTenantPage:
      type: object
      required: [items, page, size, totalItems, totalPages]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/AdminTenantView'}}
        page: {type: integer}
        size: {type: integer}
        totalItems: {type: integer, format: int64}
        totalPages: {type: integer}
    ResourceDeadLetterView:
      type: object
      required: [id, tenantId, resourceId, attempts, deadLetteredAt, createdAt, requeueCount]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        tenantId: {$ref: '#/components/schemas/UUID'}
        resourceId: {$ref: '#/components/schemas/UUID'}
        attempts: {type: integer, minimum: 5}
        lastError: {type: [string, 'null']}
        deadLetteredAt: {$ref: '#/components/schemas/Instant'}
        createdAt: {$ref: '#/components/schemas/Instant'}
        requeueCount: {type: integer, minimum: 0}
        lastRequeuedAt:
          anyOf: [{$ref: '#/components/schemas/Instant'}, {type: 'null'}]
    ResourceDeadLetterPage:
      type: object
      required: [items, page, size, totalItems, totalPages]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/ResourceDeadLetterView'}}
        page: {type: integer}
        size: {type: integer}
        totalItems: {type: integer, format: int64}
        totalPages: {type: integer}